Only create and update a framebuffer instead of using a minifb window
Instead of creating a minifb window and updating that, we now just update
the main framebuffer. This allows for more usecases where creating a
window might not be needed/wanted. This also allows for window libraries
other than minifb to be used, if you prefer.

This bumps the version to 0.3.0